refactor(pdf-service): rename /sections folder with /pdfForm in /domains/beratungshilfe/services/pdf
