+++ app/domains/beratungshilfe/services/pdf/index.ts
@@ -13,15 +13,15 @@ import { createFooter } from "~/services/pdf/footer/createFooter";
 import type { PDFDocumentBuilder } from "~/services/pdf/pdfFromUserData";
 import { pdfFromUserData } from "~/services/pdf/pdfFromUserData";
 import { createChecklistPage } from "./checklist/createChecklistPage";
-import { fillAngelegenheit } from "./sections/A_angelegenheit";
-import { fillVorraussetzungen } from "./sections/B_vorraussetzungen";
-import { fillEinkommen } from "./sections/C_einkommen";
-import { fillWohnen } from "./sections/D_wohnen";
-import { fillUnterhalt } from "./sections/E_unterhalt";
-import { fillBesitz } from "./sections/F_besitz/F_besitz";
-import { fillFooter } from "./sections/footer";
-import { fillAusgaben } from "./sections/G_ausgaben";
-import { fillHeader } from "./sections/header";
+import { fillAngelegenheit } from "./pdfForm/A_angelegenheit";
+import { fillVorraussetzungen } from "./pdfForm/B_vorraussetzungen";
+import { fillEinkommen } from "./pdfForm/C_einkommen";
+import { fillWohnen } from "./pdfForm/D_wohnen";
+import { fillUnterhalt } from "./pdfForm/E_unterhalt";
+import { fillBesitz } from "./pdfForm/F_besitz/F_besitz";
+import { fillFooter } from "./pdfForm/footer";
+import { fillAusgaben } from "./pdfForm/G_ausgaben";
+import { fillHeader } from "./pdfForm/header";
 export { getBeratungshilfeParameters };
 
 export type BerHPdfFillFunction = PdfFillFunction<
