+++ app/domains/beratungshilfe/services/pdf/pdfForm/footer.ts
