+++ app/domains/prozesskostenhilfe/services/pdf/J_belastungen.ts
@@ -1,4 +1,4 @@
-import { ausgabenSituationMapping } from "~/domains/beratungshilfe/services/pdf/sections/G_ausgaben";
+import { ausgabenSituationMapping } from "~/domains/beratungshilfe/services/pdf/pdfForm/G_ausgaben";
 import { finanzielleAngabeEinkuenfteGuards as einkuenfteGuards } from "~/domains/prozesskostenhilfe/formular/finanzielleAngaben/einkuenfte/guards";
 import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/services/pdf/attachment";
 import type { PkhPdfFillFunction } from ".";
